cpg_infra/billing_aggregator/update_budget/requirements.txt
flask
functions-framework
google-cloud-bigquery==3.13.0
google-cloud-billing-budgets==1.12.1